Rename app folder to example
The `app` folder contains an example of how to use `flutter_quill` and
should therefore be called `example`. But we already have an `example`
folder. Since the code in the existing `example` folder is already in
the `README.md`, I suggest we remove the old `example` folder and
rename `app` to `example`. Another advantage is that `flutter_quill`
becomes consistent with all other packages, since it is very unusual
to have an `app` folder.