Fix image button cancel causes crash
The PR closes the issue 137. I also restructured the code slightly
to make it more readable and fixed some minor bugs in the desktop
and web version.

I also noticed that the image button doesn't work in the desktop
version as well as in the web version. This is not part of the PR.