Hide implementation files
Hiding the implementation files is best practice according to Dart
organising a library package, https://bit.ly/3fA23Yz. It also helps
us to know which files we can change without worrying about breaking
changes.

The PR is not a breaking change, because the classes can still be
imported via the old files. In the next breaking change update, these
old files can then be removed.

No functionality was changed in the PR.