Request keyboard focus when no child is found
`getRenderEditor()!.selectWordEdge` and
`getRenderEditor().selectPosition` currently throw a `'No child'` which
prevents the field from focusing. This happens when `minHeight` is set
and the whole field is not filled with content from the editor (e. g.
it's empty or only one line has some data).

This change catches the `'No child'` exception and lets the code to
continue its execution and request field focus.

This was fixed in #531 but the code got changed and focusing after
exception got removed.