Add support for html named colors
Quill deltas often contain named colors from the standard HTML color palette. This is especially true if the delta was created using Quill on web.

These named colors could not be handled by flutter_quill, and so an exception was thrown.

This change adds dart contants for the 140 named html colors. It modifies `stringToColor` to check if the input is one of the named colors, and returns the value if found. Web colors are case insensitive.